+++ app/Http/Controllers/Admin/SettingController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers\Admin;
 
-use Gate;
 use Session;
+use Auth;
 use App\Logs;
 use App\Setting;
 use Illuminate\Http\Request;
@@ -20,7 +20,7 @@ class SettingController extends Controller
      */
     public function index()
     {
-        abort_unless(Gate::allows('setting_access'), 403);
+        abort_unless(\Gate::allows('setting_access') || Auth::user()->is_superadmin, 403);
 
         $settings = Setting::all();
 
@@ -34,7 +34,7 @@ class SettingController extends Controller
      */
     public function create()
     {
-        abort_unless(\Gate::allows('setting_create'), 403);
+        abort_unless(\Gate::allows('setting_create') || Auth::user()->is_superadmin, 403);
 
         return view('admin.settings.create');
     }
@@ -47,7 +47,7 @@ class SettingController extends Controller
      */
     public function store(StoreSettingRequest $request)
     {
-        abort_unless(\Gate::allows('setting_create'), 403);
+        abort_unless(\Gate::allows('setting_create') || Auth::user()->is_superadmin, 403);
 
         try {
             $setting = Setting::create($request->all());
@@ -74,7 +74,7 @@ class SettingController extends Controller
      */
     public function show(Setting $setting)
     {
-        abort_unless(\Gate::allows('setting_show'), 403);
+        abort_unless(\Gate::allows('setting_show') || Auth::user()->is_superadmin, 403);
 
         Logs::registerLog('Visualizou os detalhes de um parâmetro do sistema.');
 
@@ -89,7 +89,7 @@ class SettingController extends Controller
      */
     public function edit(Setting $setting)
     {
-        abort_unless(\Gate::allows('setting_edit'), 403);
+        abort_unless(\Gate::allows('setting_edit') || Auth::user()->is_superadmin, 403);
 
         $this->saveSetting($setting);
 
@@ -105,7 +105,7 @@ class SettingController extends Controller
      */
     public function update(UpdateSettingRequest $request, Setting $setting)
     {
-        abort_unless(\Gate::allows('setting_edit'), 403);
+        abort_unless(\Gate::allows('setting_edit') || Auth::user()->is_superadmin, 403);
 
         try {
             $setting->update($request->all());
@@ -132,7 +132,7 @@ class SettingController extends Controller
      */
     public function destroy($id)
     {
-        abort_unless(\Gate::allows('setting_delete'), 403);
+        abort_unless(\Gate::allows('setting_delete') || Auth::user()->is_superadmin, 403);
 
         try {
             Setting::where('id', $id)->delete();
@@ -148,7 +148,7 @@ class SettingController extends Controller
 
     public function getContent()
     {
-        abort_unless(\Gate::allows('setting_content'), 403);
+        abort_unless(\Gate::allows('setting_content') || Auth::user()->is_superadmin, 403);
 
         $settings = Setting::all();
         $this->saveSetting($settings);
@@ -158,7 +158,7 @@ class SettingController extends Controller
 
     public function saveContent(Request $request)
     {
-        abort_unless(\Gate::allows('setting_content'), 403);
+        abort_unless(\Gate::allows('setting_content') || Auth::user()->is_superadmin, 403);
 
         $values = $request->all();
